third_party/blink/web_tests/wpt_internal/fenced_frame/revoke-manual-report-event-beacons.https.html

<!DOCTYPE html>
<title>Test that window.fence.reportEvent() fails after disableUntrustedNetwork().</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>

<body>
  <script>
  promise_test(async(t) => {
    const fencedframe = await attachFencedFrameContext({generator_api: 'fledge',
                                                        register_beacon: true});
    // This page will call reportEvent twice: Once for an enum event, and once
    // for a custom URL event. Both beacons are expected not to send, because
    // network revocation should prevent the requests from going out.
    await fencedframe.execute(async () => {

      await window.fence.disableUntrustedNetwork();

      const destination_enum_event = {
        eventType: 'click',
        eventData: 'enum',
        destination: ['buyer']
      }
      window.fence.reportEvent(destination_enum_event);

      const destination_url = new URL(BEACON_URL + "?type=url",
          get_host_info().HTTPS_ORIGIN);

      const destination_url_event = {
        destinationURL: destination_url
      }
      window.fence.reportEvent(destination_url_event);
    });

  const enum_promise = nextBeacon('click', 'enum');
  const enum_result = await Promise.race([
      enum_promise,
      new Promise((resolve, reject) => t.step_timeout(
        () => resolve('timeout'), 2000))]);
  assert_equals(enum_result, 'timeout');

  const url_promise = nextBeacon('url', '<No data>');
  const url_result = await Promise.race([
      url_promise,
      new Promise((resolve, reject) => t.step_timeout(
        () => resolve('timeout'), 2000))]);
  assert_equals(url_result, 'timeout');

  }, 'Test that window.fence.reportEvent() fails after disableUntrustedNetwork().');
  </script>
  </body>
